test: extract MockDebugBundleExporter.swift from ExportTestSupport.swift
Splits the `MockDebugBundleExporter` test double out of
ExportTestSupport.swift into its own file, leaving only
MockClipboardService in the shared support file.

Byte-identical move of lines 57-66. Both files retain their test imports.

ExportTestSupport.swift: 98 → 86 lines. Tests: 667.
